+++ src/reporter.ts
@@ -91,7 +91,7 @@ export class RPReporter implements Reporter {
     this.config = {
       uploadTrace: true,
       uploadVideo: true,
-      reportLastErrorLog: true,
+      extendTestDescriptionWithLastError: true,
       ...config,
       launchId: process.env.RP_LAUNCH_ID || config.launchId,
     };
@@ -488,7 +488,7 @@ export class RPReporter implements Reporter {
       });
     }
 
-    if (result.error && this.config.reportLastErrorLog) {
+    if (result.error && this.config.extendTestDescriptionWithLastError) {
       const stacktrace = stripAnsi(result.error.stack || result.error.message);
       this.sendLog(testItemId, {
         level: LOG_LEVELS.ERROR,