+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added
 - `ReportingApi` from `@reportportal/agent-js-playwright/promises` methods (***addAttributes, setDescription, setTestCaseId, setStatus***, and all methods for setting custom statuses for test or suite) now using ***testInfo.attach*** method to attach custom data to test case.
-- `reportLastErrorLog` option to the RP config to have the ability to attach the last error log
+- `extendTestDescriptionWithLastError` option to the RP config to have the ability to attach the last error log
 
 ## [5.1.4] - 2023-10-05
 ## Changed